requirements.txt
img2pdf==0.4.4
Pillow==9.1.0
PyPDF2==2.8.1